refactorisation DONE
