--- src/scicomp/eig_val_calc/__init__.py
@@ -1,5 +1,0 @@
-"""__init__.py file."""
-
-from .equation_solver_components.solving_equation import (
-    select_solver as select_solver,
-)
