--- src/scicomp/eig_val_calc/equation_solver_components/__init__.py
@@ -1,1 +1,0 @@
-"""__init__.py file."""
