+++ src/scicomp/model_elements/solvers.py
