--- tests/test_example.py
@@ -1,3 +1,0 @@
-def test_example():
-    """Only included to make sure GitHub Action succeeds despite no real tests."""
-    assert 1 + 1 == 2
